+++ internal/provider/networking/network_ipspace_resource.go
@@ -1,0 +1,288 @@
+package networking
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	"github.com/netapp/terraform-provider-netapp-ontap/internal/provider/connection"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/netapp/terraform-provider-netapp-ontap/internal/interfaces"
+	"github.com/netapp/terraform-provider-netapp-ontap/internal/utils"
+)
+
+// Ensure provider defined types fully satisfy framework interfaces
+var _ resource.Resource = &IPspaceResource{}
+var _ resource.ResourceWithImportState = &IPspaceResource{}
+
+// NewIPspaceResource is a helper function to simplify the provider implementation.
+func NewIPspaceResource() resource.Resource {
+	return &IPspaceResource{
+		config: connection.ResourceOrDataSourceConfig{
+			Name: "network_ipspace",
+		},
+	}
+}
+
+// IPspaceResource defines the resource implementation.
+type IPspaceResource struct {
+	config connection.ResourceOrDataSourceConfig
+}
+
+// IPspaceResourceModel describes the resource data model.
+type IPspaceResourceModel struct {
+	CxProfileName  types.String  `tfsdk:"cx_profile_name"`
+	Name           types.String  `tfsdk:"name"`
+	ID             types.String  `tfsdk:"id"`
+}
+
+// Metadata returns the resource type name
+func (r *IPspaceResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_" + r.config.Name
+}
+
+// Schema defines the schema for the resource.
+func (r *IPspaceResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		// This description is used by the documentation generator and the language server.
+		MarkdownDescription: "IPspace resource",
+		Attributes: map[string]schema.Attribute{
+			"cx_profile_name": schema.StringAttribute{
+				MarkdownDescription: "Connection profile name",
+				Required:            true,
+			},
+			"name": schema.StringAttribute{
+				MarkdownDescription: "Specifies the name of the IPspace.",
+				Required:            true,
+			},
+			"id": schema.StringAttribute{
+				MarkdownDescription: "The UUID of the IPspace.",
+				Computed: true,
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.UseStateForUnknown(),
+				},
+			},
+		},
+	}
+}
+
+// Configure adds the provider configured client to the resource.
+func (r *IPspaceResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+	// Prevent panic if the provider has not been configured.
+	if req.ProviderData == nil {
+		return
+	}
+	config, ok := req.ProviderData.(connection.Config)
+	if !ok {
+		resp.Diagnostics.AddError(
+			"Unexpected Resource Configure Type",
+			fmt.Sprintf("Expected Config, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+		)
+	}
+	r.config.ProviderConfig = config
+}
+
+// Read refreshes the Terraform state with the latest data.
+func (r *IPspaceResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var data IPspaceResourceModel
+
+	// Read Terraform prior state data in to the model
+	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	errorHandler := utils.NewErrorHandler(ctx, &resp.Diagnostics)
+	// we need to defer setting the client until we can read the connection profile name
+	client, err := connection.GetRestClient(errorHandler, r.config, data.CxProfileName)
+	if err != nil {
+		// error reporting done inside New Client
+		return
+	}
+
+	cluster, err := interfaces.GetCluster(errorHandler, *client)
+	if err != nil {
+		// error reporting done inside GetCluster
+		return
+	}
+	if cluster == nil {
+		errorHandler.MakeAndReportError("No cluster found", "cluster not found")
+		return
+	}
+
+	var restInfo *interfaces.IPspaceGetDataModelONTAP
+	if !data.ID.IsNull() && !data.ID.IsUnknown() && strings.TrimSpace(data.ID.ValueString()) != "" {
+		restInfo, err = interfaces.GetIPspaceByUUID(errorHandler, *client, data.ID.ValueString())
+		if err != nil {
+			// error reporting done inside GetIPspaceByUUID
+			return
+		}
+	} else {
+		restInfo, err = interfaces.GetIPspace(errorHandler, *client, data.Name.ValueString())
+		if err != nil {
+			// error reporting done inside GetIPspace
+			return
+		}
+	}
+
+	data.ID = types.StringValue(restInfo.UUID)
+	data.Name = types.StringValue(restInfo.Name)
+
+	// Write logs using the tflog package
+	// Documentation: https://terraform.io/plugin/log
+	tflog.Debug(ctx, fmt.Sprintf("read a resource: %#v", data))
+
+	// Save data into Terraform state
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+// Create a resource and retrieve ID
+func (r *IPspaceResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var data IPspaceResourceModel
+
+	// Read Terraform plan data into the model.
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
+
+	var body interfaces.IPspaceResourceBodyDataModel
+	errorHandler := utils.NewErrorHandler(ctx, &resp.Diagnostics)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	client, err := connection.GetRestClient(errorHandler, r.config, data.CxProfileName)
+	if err != nil {
+		// error reporting done inside NewClient
+		return
+	}
+
+	cluster, err := interfaces.GetCluster(errorHandler, *client)
+	if err != nil {
+		// error reporting done inside GetCluster
+		return
+	}
+	if cluster == nil {
+		errorHandler.MakeAndReportError("No cluster found", "cluster not found")
+		return
+	}
+
+	nameValue := data.Name.ValueString()
+	body.Name = &nameValue
+
+	resource, err := interfaces.CreateIPspace(errorHandler, *client, body)
+	if err != nil {
+		// error reporting done inside CreateIPspace
+		return
+	}
+	data.ID = types.StringValue(resource.UUID)
+	data.Name = types.StringValue(resource.Name)
+
+	tflog.Trace(ctx, fmt.Sprintf("created IPspace resource, ID=%s", data.ID.ValueString()))
+
+	// Save data into Terraform state
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+// // Update updates the resource and sets the updated Terraform state on success.
+func (r *IPspaceResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var plan, state IPspaceResourceModel
+
+	// Read Terraform plan data into the model
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+	// Read Terraform state data in to the model
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	errorHandler := utils.NewErrorHandler(ctx, &resp.Diagnostics)
+	client, err := connection.GetRestClient(errorHandler, r.config, state.CxProfileName)
+	if err != nil {
+		// error reporting done inside NewClient
+		return
+	}
+
+	if state.ID.IsNull() || state.ID.IsUnknown() || strings.TrimSpace(state.ID.ValueString()) == "" {
+		errorHandler.MakeAndReportError("ID is null", "IPspace UUID is null")
+		return
+	}
+
+	var body interfaces.IPspaceResourceBodyDataModel
+	nameValue := plan.Name.ValueString()
+	body.Name = &nameValue
+
+	err = interfaces.UpdateIPspace(errorHandler, *client, body, state.ID.ValueString())
+	if err != nil {
+		// error reporting done inside UpdateIPspace
+		return
+	}
+
+	restInfo, err := interfaces.GetIPspaceByUUID(errorHandler, *client, state.ID.ValueString())
+	if err != nil {
+		// error reporting done inside GetIPspaceByUUID
+		return
+	}
+
+	plan.ID = types.StringValue(restInfo.UUID)
+	plan.Name = types.StringValue(restInfo.Name)
+
+	tflog.Debug(ctx, fmt.Sprintf("updated IPspace resource: ID=%s", plan.ID.ValueString()))
+
+	// Save data into Terraform state
+	resp.Diagnostics.Append(resp.State.Set(ctx, &plan)...)
+}
+
+// Delete deletes the resource and removes the Terraform state on success.
+func (r *IPspaceResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var data IPspaceResourceModel
+
+	// Read Terraform prior state data into the model
+	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	errorHandler := utils.NewErrorHandler(ctx, &resp.Diagnostics)
+	client, err := connection.GetRestClient(errorHandler, r.config, data.CxProfileName)
+	if err != nil {
+		// error reporting done inside NewClient
+		return
+	}
+
+	if data.ID.IsNull() || data.ID.IsUnknown() || strings.TrimSpace(data.ID.ValueString()) == "" {
+		errorHandler.MakeAndReportError("ID is null", "IPspace UUID is null")
+		return
+	}
+
+	err = interfaces.DeleteIPspace(errorHandler, *client, data.ID.ValueString())
+	if err != nil {
+		return
+	}
+}
+
+// ImportState imports a resource using ID from terraform import command by calling the Read method.
+func (r *IPspaceResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+	idParts := strings.Split(req.ID, ",")
+
+	if len(idParts) != 2 || idParts[0] == "" || idParts[1] == "" {
+		resp.Diagnostics.AddError(
+			"Unexpected Import Identifier",
+			fmt.Sprintf("Expected import identifier with format: name,cx_profile_name. Got: %q", req.ID),
+		)
+		return
+	}
+
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("name"), idParts[0])...)
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("cx_profile_name"), idParts[1])...)
+}