+++ CHANGELOG.md
@@ -6,6 +6,8 @@ FEATURES:
 - **New Resource:** `netapp-ontap_unix_group` ([#680](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/680))
 - **New Data Source:** `netapp-ontap_nvme_namespace` and `netapp-ontap_nvme_namespaces`([#483](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/483))
 - **New Resource:** `netapp-ontap_nvme_namespace` ([#482](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/482))
+- **New Data Source:** `netapp-ontap_network_ipspace` and `netapp-ontap_network_ipspaces`([#612](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/612))
+- **New Resource:** `netapp-ontap_network_ipspace` ([#612](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/612))
 
 ENHANCEMENTS: